pyproject.toml
[project]
name = "Glinteco Showoff"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Glinteco Developers", email = "developers@glinteco.com"}
]

[tool.black]
line-length = 119
target-version = ['py312']

[tool.isort]
profile = "black"
combine-as = true
line-length = 119
src-paths = ["."]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests"
filterwarnings = ["ignore::DeprecationWarning", "ignore::UserWarning"]
addopts = ["--tb=short", "-v", "--reuse-db"]

[tool.poetry]
name = "Glinteco Showoff"
version = "0.1.0"
description = "Project for showing off capability of Glinteco Company"
authors = ["Glinteco Developers <developers@glinteco.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "5.1"
django-extensions = "^3.2.3"
django-filter = "^25.1"
dj-database-url = "^2.3.0"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.10"
redis = "^5.2.1"
easy-thumbnails = "^2.10"
sentry-sdk = "^2.24.1"

djangorestframework = "^3.15.2"
drf-standardized-errors = "^0.14.1"
djangorestframework-simplejwt = "^5.5.0"
drf-yasg = "^1.21.10"
django-cors-headers = "^4.7.0"

django-storages = {extras = ["s3"], version = "^1.14.5"}

celery = "^5.4.0"
flower = "^2.0.1"
django-celery-beat = "^2.7.0"
django-celery-results = "^2.5.1"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.3"
pytest-django = "^4.10.0"
pytest-cov = "^6.0.0"
pytest-xdist = "^3.6.1"
ipython = "^9.0.2"
pre-commit = "^4.2.0"
factory-boy = "^3.3.3"
mypy = "^1.15.0"
django-stubs = {extras = ["compatible-mypy"], version = "^5.1.3"}
django-debug-toolbar = "^5.1.0"
